+++ packages/terra-graphs/CHANGELOG.md
@@ -15,6 +15,7 @@
   * Added CSS and new example tests for timeline graph.
   * Added reflow to the Paired result component.
   * Added reflow to the Bar component.
+  * Added reflow to the Bubble component.
 
 * Changed
   * Made changes to WDIO test url's according to terra rules.